Moved merge_sort to it's own module in its own file
