+++ lib/deck.rb
@@ -1,3 +1,4 @@
+
 class Deck
   attr_reader :cards
 
@@ -25,48 +26,6 @@ class Deck
     result
   end
 
-
-
-end
-
-
-class Array
-  def merge_sort(other = nil)
-    if other
-      result = []
-      (other.size + self.size).times do
-        # require 'pry'; binding.pry
-        if self.empty?
-          return result += other
-        elsif other.empty?
-          return result += self
-        end
-
-        if self[0] < other[0]
-          result << self.shift
-        else
-          result << other.shift
-        end
-      end
-    else
-      result = self.map{|el| [el]}
-
-
-      Math.log2(result.size).ceil.times do
-        new_result = []
-        result.each_slice(2).map do |a,b|
-          if b == nil
-            new_result << a
-          else
-            new_result << a.merge_sort(b)
-          end
-        end
-
-        result = new_result
-
-      end
-    end
-    result[0]
-  end
+  include MergeSortable
 
 end
